use shared readJsonFile function
